refactor: update sql client
