+++ crates/etl-destinations/src/snowflake/schema.rs
@@ -1,9 +1,14 @@
 use etl::types::{ColumnSchema, Type, is_array_type};
 
-use crate::snowflake::{Error, Result, sql_client::quote_identifier};
+use crate::snowflake::{Error, Result};
 
 pub(crate) const CDC_OPERATION_COLUMN: &str = "_cdc_operation";
 pub(crate) const CDC_SEQUENCE_COLUMN: &str = "_cdc_sequence_number";
+
+/// Double-quote a SQL identifier, escaping internal double-quotes.
+pub(crate) fn quote_identifier(name: &str) -> String {
+    format!("\"{}\"", name.replace('"', "\"\""))
+}
 
 /// Returns the Snowflake DDL type string for a given Postgres type.
 ///
@@ -126,6 +131,15 @@ mod tests {
                 *should_err,
                 "columns: {col_names:?}"
             );
+        }
+    }
+
+    #[test]
+    fn quote_identifier_cases() {
+        let cases =
+            [("my_table", r#""my_table""#), (r#"my"table"#, r#""my""table""#), ("", r#""""#)];
+        for (input, expected) in cases {
+            assert_eq!(quote_identifier(input), expected, "input: {input:?}");
         }
     }
